Ensure ShaderBackground loads react-spring peer
